requirements.txt
amqp==1.4.9
anyjson==0.3.3
arrow==0.7.0
billiard==3.3.0.23
celery==3.1.23
click==6.6
dj-database-url==0.4.1
Django==1.9.5
django-datetime-utc==1.0.2
django-extensions==1.6.1
django-storages-redux==1.3.2
djangorestframework==3.3.3
drfdocs==0.0.9
factory-boy==2.6.1
fake-factory==0.5.7
kombu==3.0.35
Markdown==2.6.6
ordereddict==1.1
psycopg2==2.6.1
python-dateutil==2.5.2
python-dotenv==0.4.0
pytz==2016.3
six==1.10.0
whitenoise==3.0
